-memberService 단위 테스트 완료. -단, 추후에 댓글과 게시글 부분 리팩토링 한 다음에 삭제부분은 테스트 재작성이 필요. 또한 , JpaRepository를 의존관계로 사용하기 때문에 , mock객체를 이용하였지만 , 너무 무분별하게 사용한 것은 아닌가 고민이 되고 있음

+++ src/main/java/com/toy/toy/service/MemberService.java
@@ -36,11 +36,6 @@ public class MemberService {
     }
 
 
-    //회원 목록 조회
-    public Page<Member> findAll(Pageable pageable){
-        return memberRepository.findAll(pageable);
-    }
-
 
     //회원 상세 조회
     public Member findById(Long memberId){
@@ -72,12 +67,12 @@ public class MemberService {
                 .orElseThrow(() -> new MemberNotFoundException("존재하지 않는 회원입니다."));
 
         //회원이 작성한 댓글들을 삭제
-        if(!member.getComments().isEmpty()){
+        if(member.getComments()!=null && !member.getComments().isEmpty()){
             commentRepository.deleteByMember(member);
         }
 
         //회원이 작성한 게시글과 거기에 작성된 댓글,파일 삭제
-        if(!member.getBoards().isEmpty()){
+        if(member.getComments()!=null && !member.getBoards().isEmpty()){
             boardService.deleteByMember(member);
         }
